TacticsParticipant.cs Cleaned Up
Also fixed typing inconsistency. All deltas are now type double and not float.

+++ utilities/models/world/combat/participant/service/TacticsParticipantCombatService.cs
@@ -11,5 +11,5 @@ public partial class TacticsParticipantCombatService : RefCounted
         TacticsControlsResource controls
     ) { }
 
-    public void AttackPawn(float delta, bool isPlayer) { }
+    public void AttackPawn(double delta, bool isPlayer) { }
 }
